# internal/kobo/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kobo store sync API, as kobibri speaks it",
    "version": "1.0.0",
    "summary": "The endpoints a Kobo e-reader calls, and what this server does with each.",
...
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "/kobo/{token}",
      "description": "This server",
      "variables": {
        "token": {
          "default": "your-sync-token",
          "description": "From the reader's page in the web interface."
        }
      }
    }
  ],
...
  "paths": {
...
    "/v1/library/{uuid}": {
      "delete": {
        "tags": [
          "Library"
        ],
        "summary": "The reader deleted a book on the device",
        "description": "Records a tombstone scoped to this one device. The book stays in the library, stays on every other reader, and is never offered to this one again. Answers 204 either way, including for a book that was never here: the device believes it is gone regardless.",
        "operationId": "deleteEntitlement",
        "x-kobibri": "served",
        "parameters": [
          {
            "$ref": "#/components/parameters/BookUUID"
          }
        ],
        "responses": {
          "204": {
            "description": "Noted."
          }
        },
        "x-kobibri-evidence": "implemented",
        "x-kobibri-evidence-note": "Nobody has deleted a book on a device while a trace was running. The tombstone path is covered by tests, not by hardware."
      }
    },
    "/v1/library/tags": {
      "post": {
        "tags": [
          "Collections"
        ],
        "summary": "Create a shelf",
        "description": "Answers 201 with the new shelf's uuid as a bare JSON string — not an object.",
        "operationId": "createTag",
        "x-kobibri": "served",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "Name": {
                    "type": "string"
                  },
                  "Items": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/TagItem"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The shelf uuid, as a bare string.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                },
                "example": "\"4d05cc0c-8151-3fc8-8b2b-7c41afe002f0\""
              }
            }
          }
        },
        "x-kobibri-evidence": "implemented",
        "x-kobibri-evidence-note": "Collections are covered by tests and by the fake device, not by a trace from real hardware."
      },
      "delete": {
        "tags": [
          "Collections"
        ],
        "summary": "Refused on purpose",
        "description": "A guard, not a feature. Without it `DELETE /v1/library/tags` matches the book-deletion route and is read as deleting a book called \"tags\".",
        "operationId": "deleteTagsGuard",
        "x-kobibri": "served",
        "responses": {
          "405": {
            "description": "Method not allowed."
          }
        },
        "x-kobibri-evidence": "implemented",
        "x-kobibri-evidence-note": "A guard against a routing collision. No device is expected to send it."
      }
    },
    "/v1/library/tags/{id}": {
      "put": {
        "tags": [
          "Collections"
        ],
        "summary": "Rename a shelf",
        "description": "Shares its path shape with reading progress — `PUT /v1/library/X/Y` is both — and no router can tell them apart, so one handler does it: a first segment of `tags` can only be a collection, since a book id is always a uuid.",
        "operationId": "renameTag",
        "x-kobibri": "served",
        "parameters": [
          {
            "$ref": "#/components/parameters/TagID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "Name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Renamed."
          }
        },
        "x-kobibri-evidence": "implemented",
        "x-kobibri-evidence-note": "Tests only."
      },
      "delete": {
        "tags": [
          "Collections"
        ],
        "summary": "Delete a shelf",
        "description": "Soft-deleted here. Without the remembered row the next sync cannot send `DeletedTag`, and the shelf sticks on every other device forever — you can only delete what you still remember.",
        "operationId": "deleteTag",
        "x-kobibri": "served",
        "parameters": [
          {
            "$ref": "#/components/parameters/TagID"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted."
          }
        },
        "x-kobibri-evidence": "implemented",
        "x-kobibri-evidence-note": "Tests only."
      }
    },
...
    "/v1/affiliate": {
      "get": {
        "tags": [
          "Store"
        ],
        "summary": "Affiliate",
        "description": "Relayed, and the store always refuses: this is one of the two endpoints that check the access token's version, and the reader is holding one this server invented. `400 Invalid token version`, every sync, permanently.\n\nRelaying that 400 to the device is what aborted the whole sync — the reader logged `FailedSync reason=WebRequestErr` and threw away a library sync that had just succeeded. The refusal stops here and `{\"Affiliate\":\"Kobo\"}` goes out instead.",
        "operationId": "affiliate",
        "x-kobibri": "relayed",
        "parameters": [
          {
            "name": "PlatformID",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SerialNumber",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The affiliate name."
          }
        },
        "x-kobibri-evidence": "observed",
        "x-kobibri-evidence-note": "Kobo Libra Colour, fw 4.45.23697. Refused by the store with 400 Invalid token version on every single sync."
      }
    }
  },
  "components": {
    "parameters": {
      "BookUUID": {
        "name": "uuid",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "description": "The canonical book id. Issued once and never reissued — it is the only identifier the device knows."
      },
      "TagID": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "description": "The shelf uuid."
      }
    },
    "schemas": {
...
      "TagItem": {
        "type": "object",
        "properties": {
          "RevisionId": {
            "type": "string",
            "format": "uuid"
          },
          "Type": {
            "type": "string",
            "enum": [
              "ProductRevisionTagItem"
            ]
          }
        }
      },
...
    }
  }
}